+++ jans-config-api/server/src/main/java/io/jans/configapi/rest/resource/auth/AttributesResource.java
@@ -16,7 +16,7 @@ import io.jans.configapi.util.ApiAccessConstants;
 import io.jans.configapi.util.ApiConstants;
 import io.jans.configapi.util.AttributeNames;
 import io.jans.configapi.core.util.Jackson;
-import io.jans.model.GluuAttribute;
+import io.jans.model.JansAttribute;
 import io.jans.orm.model.PagedResult;
 
 import io.swagger.v3.oas.annotations.Operation;
@@ -53,7 +53,7 @@ import org.slf4j.Logger;
 @Produces(MediaType.APPLICATION_JSON)
 public class AttributesResource extends ConfigBaseResource {
 
-    private static final String GLUU_ATTRIBUTE = "gluu attribute";
+    private static final String JANS_ATTRIBUTE = "jans attribute";
 
     @Inject
     Logger log;
@@ -61,7 +61,7 @@ public class AttributesResource extends ConfigBaseResource {
     @Inject
     AttributeService attributeService;
 
-    @Operation(summary = "Gets a list of Gluu attributes.", description = "Gets a list of Gluu attributes.", operationId = "get-attributes", tags = {
+    @Operation(summary = "Gets a list of Jans attributes.", description = "Gets a list of Jans attributes.", operationId = "get-attributes", tags = {
             "Attribute" }, security = @SecurityRequirement(name = "oauth2", scopes = {
                     ApiAccessConstants.ATTRIBUTES_READ_ACCESS }))
     @ApiResponses(value = {
@@ -88,7 +88,7 @@ public class AttributesResource extends ConfigBaseResource {
         }
               
         SearchRequest searchReq = createSearchRequest(attributeService.getDnForAttribute(null), pattern, sortBy,
-                sortOrder, startIndex, limit, null, null, this.getMaxCount(),fieldValuePair, GluuAttribute.class);
+                sortOrder, startIndex, limit, null, null, this.getMaxCount(),fieldValuePair, JansAttribute.class);
 
         return Response.ok(doSearch(searchReq, status)).build();
     }
@@ -97,7 +97,7 @@ public class AttributesResource extends ConfigBaseResource {
             "Attribute" }, security = @SecurityRequirement(name = "oauth2", scopes = {
                     ApiAccessConstants.ATTRIBUTES_READ_ACCESS }))
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Ok", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = GluuAttribute.class), examples = @ExampleObject(name = "Response example", value = "example/attribute/attribute-get.json"))),
+            @ApiResponse(responseCode = "200", description = "Ok", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = JansAttribute.class), examples = @ExampleObject(name = "Response example", value = "example/attribute/attribute-get.json"))),
             @ApiResponse(responseCode = "401", description = "Unauthorized"),
             @ApiResponse(responseCode = "500", description = "InternalServerError") })
     @GET
@@ -105,24 +105,24 @@ public class AttributesResource extends ConfigBaseResource {
             ApiAccessConstants.ATTRIBUTES_WRITE_ACCESS }, superScopes = { ApiAccessConstants.SUPER_ADMIN_READ_ACCESS })
     @Path(ApiConstants.INUM_PATH)
     public Response getAttributeByInum(@Parameter(description = "Attribute Id") @PathParam(ApiConstants.INUM) @NotNull String inum) {
-        GluuAttribute attribute = attributeService.getAttributeByInum(inum);
-        checkResourceNotNull(attribute, GLUU_ATTRIBUTE);
+        JansAttribute attribute = attributeService.getAttributeByInum(inum);
+        checkResourceNotNull(attribute, JANS_ATTRIBUTE);
         return Response.ok(attribute).build();
     }
 
     @Operation(summary = "Adds a new attribute", description = "Adds a new attribute", operationId = "post-attributes", tags = {
             "Attribute" }, security = @SecurityRequirement(name = "oauth2", scopes = {
                     ApiAccessConstants.ATTRIBUTES_WRITE_ACCESS }))
-    @RequestBody(description = "GluuAttribute object", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = GluuAttribute.class), examples = @ExampleObject(name = "Request example", value = "example/attribute/attribute.json")))
+    @RequestBody(description = "JansAttribute object", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = JansAttribute.class), examples = @ExampleObject(name = "Request example", value = "example/attribute/attribute.json")))
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "201", description = "Created", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = GluuAttribute.class), examples = @ExampleObject(name = "Response example", value = "example/attribute/attribute.json"))),
+            @ApiResponse(responseCode = "201", description = "Created", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = JansAttribute.class), examples = @ExampleObject(name = "Response example", value = "example/attribute/attribute.json"))),
             @ApiResponse(responseCode = "401", description = "Unauthorized"),
             @ApiResponse(responseCode = "500", description = "InternalServerError") })
     @POST
     @ProtectedApi(scopes = { ApiAccessConstants.ATTRIBUTES_WRITE_ACCESS }, groupScopes = {}, superScopes = {
             ApiAccessConstants.SUPER_ADMIN_WRITE_ACCESS })
-    public Response createAttribute(@Valid GluuAttribute attribute) {
-        log.debug(" GluuAttribute details to add - attribute:{}", attribute);
+    public Response createAttribute(@Valid JansAttribute attribute) {
+        log.debug(" JansAttribute details to add - attribute:{}", attribute);
         checkNotNull(attribute.getName(), AttributeNames.NAME);
         checkNotNull(attribute.getDisplayName(), AttributeNames.DISPLAY_NAME);
         checkResourceNotNull(attribute.getDataType(), AttributeNames.DATA_TYPE);
@@ -130,43 +130,43 @@ public class AttributesResource extends ConfigBaseResource {
         attribute.setInum(inum);
         attribute.setDn(attributeService.getDnForAttribute(inum));
         attributeService.addAttribute(attribute);
-        GluuAttribute result = attributeService.getAttributeByInum(inum);
+        JansAttribute result = attributeService.getAttributeByInum(inum);
         return Response.status(Response.Status.CREATED).entity(result).build();
     }
 
     @Operation(summary = "Updates an existing attribute", description = "Updates an existing attribute", operationId = "put-attributes", tags = {
             "Attribute" }, security = @SecurityRequirement(name = "oauth2", scopes = {
                     ApiAccessConstants.ATTRIBUTES_WRITE_ACCESS }))
-    @RequestBody(description = "GluuAttribute object", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = GluuAttribute.class), examples = @ExampleObject(name = "Request example", value = "example/attribute/attribute.json")))
+    @RequestBody(description = "JansAttribute object", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = JansAttribute.class), examples = @ExampleObject(name = "Request example", value = "example/attribute/attribute.json")))
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Ok", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = GluuAttribute.class), examples = @ExampleObject(name = "Response example", value = "example/attribute/attribute.json"))),
+            @ApiResponse(responseCode = "200", description = "Ok", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = JansAttribute.class), examples = @ExampleObject(name = "Response example", value = "example/attribute/attribute.json"))),
             @ApiResponse(responseCode = "401", description = "Unauthorized"),
             @ApiResponse(responseCode = "500", description = "InternalServerError") })
     @PUT
     @ProtectedApi(scopes = { ApiAccessConstants.ATTRIBUTES_WRITE_ACCESS }, groupScopes = {}, superScopes = {
             ApiAccessConstants.SUPER_ADMIN_WRITE_ACCESS })
-    public Response updateAttribute(@Valid GluuAttribute attribute) {
-        log.debug(" GluuAttribute details to update - attribute:{}", attribute);
+    public Response updateAttribute(@Valid JansAttribute attribute) {
+        log.debug(" JansAttribute details to update - attribute:{}", attribute);
         String inum = attribute.getInum();
-        checkResourceNotNull(inum, GLUU_ATTRIBUTE);
+        checkResourceNotNull(inum, JANS_ATTRIBUTE);
         checkNotNull(attribute.getName(), AttributeNames.NAME);
         checkNotNull(attribute.getDisplayName(), AttributeNames.DISPLAY_NAME);
         checkResourceNotNull(attribute.getDataType(), AttributeNames.DATA_TYPE);
-        GluuAttribute existingAttribute = attributeService.getAttributeByInum(inum);
-        checkResourceNotNull(existingAttribute, GLUU_ATTRIBUTE);
+        JansAttribute existingAttribute = attributeService.getAttributeByInum(inum);
+        checkResourceNotNull(existingAttribute, JANS_ATTRIBUTE);
         attribute.setInum(existingAttribute.getInum());
         attribute.setBaseDn(attributeService.getDnForAttribute(inum));
         attributeService.updateAttribute(attribute);
-        GluuAttribute result = attributeService.getAttributeByInum(inum);
+        JansAttribute result = attributeService.getAttributeByInum(inum);
         return Response.ok(result).build();
     }
 
-    @Operation(summary = "Partially modify a GluuAttribute", description = "Partially modify a GluuAttribute", operationId = "patch-attributes-by-inum", tags = {
+    @Operation(summary = "Partially modify a JansAttribute", description = "Partially modify a JansAttribute", operationId = "patch-attributes-by-inum", tags = {
             "Attribute" }, security = @SecurityRequirement(name = "oauth2", scopes = {
                     ApiAccessConstants.ATTRIBUTES_WRITE_ACCESS }))
     @RequestBody(description = "String representing patch-document.", content = @Content(mediaType = MediaType.APPLICATION_JSON_PATCH_JSON, array = @ArraySchema(schema = @Schema(implementation = PatchRequest.class)), examples = @ExampleObject(name = "Patch request example", value = "example/attribute/attribute-patch.json")))
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Updated GluuAttribute", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = GluuAttribute.class), examples = @ExampleObject(name = "Response example", value = "example/attribute/attribute.json"))),
+            @ApiResponse(responseCode = "200", description = "Updated JansAttribute", content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = JansAttribute.class), examples = @ExampleObject(name = "Response example", value = "example/attribute/attribute.json"))),
             @ApiResponse(responseCode = "401", description = "Unauthorized"),
             @ApiResponse(responseCode = "404", description = "Not Found"),
             @ApiResponse(responseCode = "500", description = "InternalServerError") })
@@ -177,9 +177,9 @@ public class AttributesResource extends ConfigBaseResource {
     @Path(ApiConstants.INUM_PATH)
     public Response patchAtribute(@Parameter(description = "Attribute Id") @PathParam(ApiConstants.INUM) @NotNull String inum, @NotNull String pathString)
             throws JsonPatchException, IOException {
-        log.debug(" GluuAttribute details to patch - inum:{}, pathString:{}", inum, pathString);
-        GluuAttribute existingAttribute = attributeService.getAttributeByInum(inum);
-        checkResourceNotNull(existingAttribute, GLUU_ATTRIBUTE);
+        log.debug(" JansAttribute details to patch - inum:{}, pathString:{}", inum, pathString);
+        JansAttribute existingAttribute = attributeService.getAttributeByInum(inum);
+        checkResourceNotNull(existingAttribute, JANS_ATTRIBUTE);
 
         existingAttribute = Jackson.applyPatch(pathString, existingAttribute);
         attributeService.updateAttribute(existingAttribute);
@@ -198,27 +198,27 @@ public class AttributesResource extends ConfigBaseResource {
     @ProtectedApi(scopes = { ApiAccessConstants.ATTRIBUTES_DELETE_ACCESS }, groupScopes = {}, superScopes = {
             ApiAccessConstants.SUPER_ADMIN_DELETE_ACCESS })
     public Response deleteAttribute(@Parameter(description = "Attribute Id") @PathParam(ApiConstants.INUM) @NotNull String inum) {
-        log.debug(" GluuAttribute details to delete - inum:{}", inum);
-        GluuAttribute attribute = attributeService.getAttributeByInum(inum);
-        checkResourceNotNull(attribute, GLUU_ATTRIBUTE);
+        log.debug(" JansAttribute details to delete - inum:{}", inum);
+        JansAttribute attribute = attributeService.getAttributeByInum(inum);
+        checkResourceNotNull(attribute, JANS_ATTRIBUTE);
         attributeService.removeAttribute(attribute);
         return Response.noContent().build();
     }
 
-    private PagedResult<GluuAttribute> doSearch(SearchRequest searchReq, String status) {
+    private PagedResult<JansAttribute> doSearch(SearchRequest searchReq, String status) {
 
-        logger.debug("GluuAttribute search params - searchReq:{} , status:{} ", searchReq, status);
+        logger.debug("JansAttribute search params - searchReq:{} , status:{} ", searchReq, status);
 
-        PagedResult<GluuAttribute> pagedResult = attributeService.searchGluuAttributes(searchReq, status);
+        PagedResult<JansAttribute> pagedResult = attributeService.searchJansAttributes(searchReq, status);
 
         logger.debug("PagedResult  - pagedResult:{}", pagedResult);
         if (pagedResult != null) {
             logger.debug(
-                    "GluuAttributes fetched  - pagedResult.getTotalEntriesCount():{}, pagedResult.getEntriesCount():{}, pagedResult.getEntries():{}",
+                    "JansAttributes fetched  - pagedResult.getTotalEntriesCount():{}, pagedResult.getEntriesCount():{}, pagedResult.getEntries():{}",
                     pagedResult.getTotalEntriesCount(), pagedResult.getEntriesCount(), pagedResult.getEntries());
         }
 
-        logger.debug("GluuAttributes pagedResult:{} ", pagedResult);
+        logger.debug("JansAttributes pagedResult:{} ", pagedResult);
         return pagedResult;
     }