+++ jans-linux-setup/jans_setup/jans_setup.py
@@ -141,6 +141,8 @@ if base.current_app.profile == SetupProfiles.JANS or base.current_app.profile ==
     from setup_app.installers.fido import FidoInstaller
     from setup_app.installers.eleven import ElevenInstaller
     from setup_app.installers.jans_link import JansLinkInstaller
+    from setup_app.installers.casa import CasaInstaller
+
 
 from setup_app.installers.config_api import ConfigApiInstaller
 from setup_app.installers.jans_cli import JansCliInstaller
@@ -262,6 +264,7 @@ if Config.profile == SetupProfiles.JANS or Config.profile == SetupProfiles.DISA_
 
 if Config.profile == SetupProfiles.JANS:
     elevenInstaller = ElevenInstaller()
+    casa_installer = CasaInstaller()
     jans_link_installer = JansLinkInstaller()
 
 jansCliInstaller = JansCliInstaller()
@@ -446,6 +449,11 @@ def main():
                 if (Config.installed_instance and elevenInstaller.install_var in Config.addPostSetupService) or (
                         not Config.installed_instance and Config.get(elevenInstaller.install_var)):
                     elevenInstaller.start_installation()
+
+                if (Config.installed_instance and casa_installer.install_var in Config.addPostSetupService) or (
+                        not Config.installed_instance and Config.get(casa_installer.install_var)):
+                    casa_installer.start_installation()
+
 
                 if (Config.installed_instance and jans_link_installer.install_var in Config.addPostSetupService) or (
                         not Config.installed_instance and Config.get(jans_link_installer.install_var)):