+++ jans-linux-setup/jans_setup/setup_app/setup_options.py
@@ -19,6 +19,7 @@ def get_setup_options():
         'installFido2': True,
         'installEleven': False,
         'install_jans_link': True,
+        'install_casa': False,
         'loadTestData': False,
         'allowPreReleasedFeatures': False,
         'listenAllInterfaces': False,
@@ -102,6 +103,8 @@ def get_setup_options():
         if base.argsp.install_jans_link:
             setupOptions['install_jans_link'] = True
 
+        if base.argsp.with_casa:
+            setupOptions['install_casa'] = True
 
         if base.argsp.jans_max_mem:
             setupOptions['jans_max_mem'] = base.argsp.jans_max_mem