+++ jans-linux-setup/jans_setup/setup_app/utils/arg_parser.py
@@ -96,6 +96,7 @@ if PROFILE == JANS_PROFILE or PROFILE == DISA_STIG_PROFILE:
     parser.add_argument('--no-scim', help="Do not install Scim Server", action='store_true')
     parser.add_argument('--no-fido2', help="Do not install Fido2 Server", action='store_true')
     parser.add_argument('--install-jans-link', help="Install Link Server", action='store_true')
+    parser.add_argument('--with-casa', help="Install Gluu/Flex Casa Server", action='store_true')
 
     #parser.add_argument('--oxd-use-jans-storage', help="Use Jans Storage for Oxd Server", action='store_true')
     parser.add_argument('--load-config-api-test', help="Load Config Api Test Data", action='store_true')